Split Home component responsibilities
